backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
blinker==1.9.0
cffi==1.17.1
click==8.1.8
cryptography==44.0.1
ecdsa==0.19.0
fastapi==0.115.8
Flask==3.1.0
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
python-jose[cryptography]==3.3.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
typing_extensions==4.12.2
uvicorn==0.34.0
Werkzeug==3.1.3
edgedb==1.0.0
bcrypt==4.0.1
python-multipart==0.0.6
email-validator==2.0.0.post2
python-dotenv==0.19.0
pytest==7.4.4
pytest-asyncio==0.23.5
httpx==0.26.0
